+++ src/main/java/net/astr0/astech/block/ReactionChamber/ChemicalReactorBlock.java
@@ -101,7 +101,7 @@ public class ChemicalReactorBlock extends HorizontalDirectionalBlock implements 
     public void animateTick(BlockState pState, Level pLevel, BlockPos pPos, RandomSource pRandom) {
         if (pState.getValue(ModBlocks.BLOCKSTATE_ACTIVE)) {
             double d0 = (double)pPos.getX() + 0.5D;
-            double d1 = (double)pPos.getY();
+            double d1 = pPos.getY();
             double d2 = (double)pPos.getZ() + 0.5D;
 
             Direction direction = pState.getValue(FACING);
